Refactoring: move CompilationError, etc.
